Move ProjectForNamespaceFromLister from plugin/pkg/utils to pkg/utils/gardener
